requirements.txt
pydantic
pyaml
mypy
pre-commit
